:hammer: Move DesktopToastManager implementation to commonMain to support multiplatform
