+++ app/src/commonMain/kotlin/com/crosspaste/notification/ToastManager.kt
@@ -1,12 +1,32 @@
 package com.crosspaste.notification
 
+import com.crosspaste.utils.mainDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
 
-interface ToastManager {
+class ToastManager {
 
-    val toastList: StateFlow<List<Toast>>
+    private val coroutineScope = CoroutineScope(mainDispatcher)
 
-    fun pushToast(toast: Toast)
+    private val _toastList: MutableStateFlow<List<Toast>> = MutableStateFlow(listOf())
 
-    fun removeToast(messageId: Int)
+    val toastList: StateFlow<List<Toast>> = _toastList
+
+    fun pushToast(toast: Toast) {
+        this._toastList.value = listOf(toast) + this._toastList.value
+
+        toast.duration?.let { duration ->
+            coroutineScope.launch {
+                delay(duration)
+                removeToast(toast.messageId)
+            }
+        }
+    }
+
+    fun removeToast(messageId: Int) {
+        this._toastList.value = this._toastList.value.filter { it.messageId != messageId }
+    }
 }
